requirements.txt
anyio==3.6.1
asgiref==3.5.1
beautifulsoup4==4.11.1
boto3==1.23.0
botocore==1.26.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
elasticsearch==7.10.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
jmespath==1.0.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-aws4auth==1.1.2
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
